requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
opencv-python-headless
ultralytics
numpy
pillow
httpx
python-dotenv